Moved Member API filter class to a separate file.
